local stash

+++ core/src/main/java/dev/olog/core/interactor/playlist/AddToPlaylistUseCase.kt
@@ -16,20 +16,20 @@ class AddToPlaylistUseCase @Inject constructor(
 
     suspend operator fun invoke(playlist: Playlist, mediaId: MediaId) {
         if (mediaId.isLeaf && mediaId.isPodcast) {
-            podcastPlaylistGateway.addSongsToPlaylist(playlist.id, listOf(mediaId.resolveId))
+            podcastPlaylistGateway.addSongsToPlaylist(playlist.id.toLong(), listOf(mediaId.resolveId))
             return
         }
 
         if (mediaId.isLeaf) {
-            playlistGateway.addSongsToPlaylist(playlist.id, listOf(mediaId.resolveId))
+            playlistGateway.addSongsToPlaylist(playlist.id.toLong(), listOf(mediaId.resolveId))
             return
         }
 
         val songList = getSongListByParamUseCase(mediaId).map { it.id }
         if (mediaId.isAnyPodcast) {
-            podcastPlaylistGateway.addSongsToPlaylist(playlist.id, songList)
+            podcastPlaylistGateway.addSongsToPlaylist(playlist.id.toLong(), songList)
         } else {
-            playlistGateway.addSongsToPlaylist(playlist.id, songList)
+            playlistGateway.addSongsToPlaylist(playlist.id.toLong(), songList)
         }
     }
 }
